requirements.txt
fastapi
boto3
python-multipart
python-dotenv
google-generativeai
pymupdf
uvicorn